requirements.txt
beautifulsoup4==4.12.3
psycopg2_binary==2.9.10
python-dotenv==1.0.1
PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.3
streamlit==1.33.0
toml==0.10.2
